build(main): move project source to src folder
