+++ next.config.js
@@ -2,6 +2,6 @@ const path = require('path')
 
 module.exports = {
   sassOptions: {
-    includePaths: [path.join(__dirname, 'styles')],
+    includePaths: [path.join(__dirname, 'src/scss')],
   },
 }
