+++ src/components/movie/list/list.module.scss
