+++ src/components/movie/poster/poster.module.scss
