+++ src/pages/_app.tsx
@@ -1,5 +1,5 @@
 import '../styles/app.scss'
-import { Layout } from '~/components/layout/layout.module'
+import { Layout } from '~/src/components/layout/layout.module'
 import { AppProps } from 'next/app';
 
 const App: React.FC<AppProps> = ({ Component, pageProps }) => {
