+++ src/pages/movie/detail/detail.module.scss
