+++ src/pages/movie/detail/index.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import { myObserver } from '~/utils/resize-observe';
+import { myObserver } from '~/src/utils/resize-observe';
 
 export default class Detail extends React.Component {
 
-  componentWiMount(): void {
+  componentDidMount(): void {
     const someEl = document.querySelector('body');
     myObserver((entry) => console.log('resize:', entry)).observe(someEl);
   }
