+++ src/pages/movie/list/list.module.scss
